# BLPS/src/main/java/xyz/anomatver/blps/auth/dto/LoginDTO.java
package xyz.anomatver.blps.auth.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class LoginDTO {
    private String username;
    private String password;

}